requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.7
distlib==0.3.8
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-paypal==2.1
django-select2==8.1.2
djangorestframework==3.15.1
drf-extra-fields==3.7.0
filelock==3.14.0
filetype==1.2.0
identify==2.5.36
idna==3.7
mysqlclient==2.2.4
nodeenv==1.8.0
paypalrestsdk==1.13.3
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
pycparser==2.22
pyOpenSSL==24.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
virtualenv==20.26.2
